Conversion correction.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/trunk@4057 0a995953-517f-4c62-b348-104708d926b2

+++ Kernel/spot-on-neighbor.cc
@@ -1445,7 +1445,7 @@ void spoton_neighbor::slotReadyRead(void)
   if(!data.isEmpty())
     {
       QReadLocker locker1(&m_maximumBufferSizeMutex);
-      int maximumBufferSize = m_maximumBufferSize;
+      qint64 maximumBufferSize = m_maximumBufferSize;
 
       locker1.unlock();
